+++ checkingtime/.vscode/settings.json
@@ -1,15 +1,19 @@
 {
    "cSpell.words": [
       "aquare",
+      "công",
       "điện",
       "kengang",
       "khẩu",
       "midu",
       "ngang",
+      "nhân",
       "Nhat",
       "Sigin",
       "Thay",
       "thiết",
-      "thoại"
+      "thoại",
+      "Thông",
+      "việc"
    ]
 }
