+++ checkingtime/App.tsx
@@ -9,8 +9,9 @@ import HappyGames from "./src/Screens/HappyGame/HappyGames";
 import FormCreated from "./src/Screens/FormCreated/index";
 import Form from "./src/Screens/Form/Form";
 
-import SignInScreen from "./src/Screens/SignInScreen/SignInScreen";
+
 import ForgotPasswordScreen from "./src/Screens/ForgotPasswordScreen/ForgotPasswordScreen";
+import SignInScreen from "./src/Screens/SignInScreen/SignInScreen";
 
 import Navigation from "./src/navigation";
 import NhatKyDiem from "./src/Screens/NhatKyDiem/NhatKyDiem";
@@ -18,12 +19,14 @@ import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import TabsNavigation from "./src/navigation/Tabs";
 import Tabs from "./src/navigation/Tabs";
 import HomeScreen from "./src/Screens/HomeScreen/index";
+import Account from "./src/Screens/Account/Account";
+import NavigationA from "./src/navigation";
 const App = () => {
   return (
     <SafeAreaView style={styles.root}>
-      <NavigationContainer>
-      <SignInScreen />
-      </NavigationContainer>
+      
+      <NavigationA />
+      
     </SafeAreaView>
   );
 };
