+++ checkingtime/package.json
@@ -46,6 +46,7 @@
     "@types/react": "~17.0.21",
     "@types/react-native": "~0.66.13",
     "@types/react-native-vector-icons": "^6.4.11",
+    "@types/react-navigation": "^3.4.0",
     "typescript": "~4.3.5"
   },
   "private": true
