+++ checkingtime/src/Screens/Account/style.tsx
@@ -1,13 +1,15 @@
-import { StyleSheet } from "react-native";
+import { StyleSheet, StatusBar } from "react-native";
 
 const createStyles = () =>
   StyleSheet.create({
     container: {
       flex: 1,
       flexDirection: "column",
-      alignSelf: "flex-start",
-      marginTop: "20%",
-      margin: 30,
+      //alignSelf: "flex-start",
+      marginTop: 10,
+      marginLeft: 10,
+      // margin: 30,
+      paddingTop: StatusBar.currentHeight,
     },
     user: {
       fontSize: 14,
